requirements.txt
flask-restx==1.3.0
sqlalchemy==2.0.25